tests: Introduce an abstract base class for EqualityExtractorTest
Since `ColumnsUnderNotPredicateFinderTest` was extending the
`EqualityExtrractorTest`, it was also rerunning all test methods from
it. Introduce an abstract base class to share the preparation variables
and params and use it as superclass for both classes.
